pyproject.toml
[tool.poetry]
name = "shift-tasks"
version = "0.1.0"
description = ""
authors = ["Dddarknight <9801677@gmail.com>"]
readme = "README.md"
packages = [{include = "shift_tasks"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.28"
python-dotenv = "^1.0.1"
dependency-injector = "^4.41.0"
uvicorn = "^0.28.0"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
pytest-asyncio = "^0.23.5.post1"
httpx = "^0.27.0"
mimesis = "^15.1.0"
pytest-cov = "^4.1.0"
gunicorn = "^21.2.0"
ruff = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
